requirements.txt
openai>=1.12.0
azure-search-documents>=11.4.0
azure-ai-contentsafety>=1.0.0
azure-core>=1.29.0
python-dotenv>=1.0.0
